+++ example/db/schema.rb
@@ -17,6 +17,7 @@ ActiveRecord::Schema.define(version: 2018_10_29_133756) do
     t.string "name", null: false
     t.string "phone"
     t.string "address"
+    t.string "code", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end