+++ lib/database_consistency/database_processor.rb
@@ -1,0 +1,18 @@
+module DatabaseConsistency
+  # The class to process missing validators
+  class DatabaseProcessor
+    VERIFIERS = [
+      ColumnVerifiers::PresenceMissingVerifier
+    ].freeze
+
+    def reports
+      Helper.models.flat_map do |model|
+        model.columns.flat_map do |column|
+          VERIFIERS.map do |verifier|
+            verifier.verify(model, column)
+          end
+        end
+      end.compact
+    end
+  end
+end