+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
+### [0.2.0] - 2018/10/31
+
+Features:
+
+- Add check of missing presence validator
+
 ### [0.1.0] - 2018/10/30
 
 Features:
 
-- Add check of presence validator
+- Add check of consistency between presence validator and database field